rename ipfssync -> ipfsconnect imports
