+++ service/storage.go
@@ -9,7 +9,7 @@ import (
 	"go.vocdoni.io/dvote/config"
 	"go.vocdoni.io/dvote/crypto/ethereum"
 	"go.vocdoni.io/dvote/data"
-	"go.vocdoni.io/dvote/ipfssync"
+	"go.vocdoni.io/dvote/ipfsconnect"
 	"go.vocdoni.io/dvote/log"
 	"go.vocdoni.io/dvote/metrics"
 )
@@ -17,7 +17,6 @@ import (
 func IPFS(ipfsconfig *config.IPFSCfg, signer *ethereum.SignKeys,
 	ma *metrics.Agent) (storage data.Storage, err error) {
 	log.Info("creating ipfs service")
-	var storageSync ipfssync.IPFSsync
 	if !ipfsconfig.NoInit {
 		os.Setenv("IPFS_FD_MAX", "1024")
 		ipfsStore := data.IPFSNewConfig(ipfsconfig.ConfigPath)
@@ -44,19 +43,18 @@ func IPFS(ipfsconfig *config.IPFSCfg, signer *ethereum.SignKeys,
 		if len(ipfsconfig.SyncKey) > 0 {
 			log.Info("enabling ipfs synchronization")
 			_, priv := signer.HexString()
-			storageSync = *ipfssync.NewIPFSsync(
+			ipfsconn := ipfsconnect.New(
 				filepath.Join(ipfsconfig.ConfigPath, "ipfsSync"),
 				ipfsconfig.SyncKey,
 				priv,
 				"libp2p",
 				storage,
 			)
-			storageSync.OnlyConnect = true // only connect to peers, do not sync files
 			if len(ipfsconfig.SyncPeers) > 0 && len(ipfsconfig.SyncPeers[0]) > 8 {
 				log.Debugf("using custom ipfs sync bootnodes %s", ipfsconfig.SyncPeers)
-				storageSync.Transport.BootNodes = ipfsconfig.SyncPeers
+				ipfsconn.Transport.BootNodes = ipfsconfig.SyncPeers
 			}
-			storageSync.Start()
+			ipfsconn.Start()
 		}
 	}
 	return
